Int labels

+++ replibert/replibert/finetuning/evaluate.py
@@ -58,7 +58,7 @@ def _calculate_loss(model: torch.nn.Module, test_loader: DataLoader, criterion: 
         for inputs, labels in tqdm(test_loader, desc="Evaluating", unit="batch"):
             input_ids = inputs[:, 0, :].to(device, non_blocking=True)
             attention_mask = inputs[:, 1, :].to(device, non_blocking=True)
-            labels = labels.to(device).long()
+            labels = labels.to(device)
 
             logits = model(input_ids=input_ids, attention_mask=attention_mask).squeeze(-1)
             probabilities = torch.sigmoid(logits)
@@ -70,7 +70,7 @@ def _calculate_loss(model: torch.nn.Module, test_loader: DataLoader, criterion: 
 
             # Store predictions and labels for metrics
             all_predictions.extend(predictions.cpu().numpy())
-            all_labels.extend(labels.cpu().numpy())
+            all_labels.extend(labels.cpu().numpy().astype(int))
             all_probabilities.extend(probabilities.cpu().numpy())
 
     return total_loss, total_samples, all_labels, all_predictions, all_probabilities
